Move changes into `drafts` directory
