+++ src/drafts/MarkdownEditor/MarkdownEditor.tsx
@@ -1,15 +1,15 @@
 import {isMacOS} from '@primer/behaviors/utils'
 import {useSSRSafeId} from '@react-aria/ssr'
 import React, {forwardRef, useCallback, useEffect, useImperativeHandle, useMemo, useRef, useState} from 'react'
-import Box from '../Box'
-import {FileType} from '../hooks'
+import Box from '../../Box'
+import {FileType} from '../hooks/useUnifiedFileSelect'
 import {useIgnoreKeyboardActionsWhileComposing} from '../hooks/useIgnoreKeyboardActionsWhileComposing'
-import {useResizeObserver} from '../hooks/useResizeObserver'
-import {useSyntheticChange} from '../drafts/hooks/useSyntheticChange'
+import {useResizeObserver} from '../../hooks/useResizeObserver'
+import {useSyntheticChange} from '../hooks/useSyntheticChange'
 import MarkdownViewer from '../MarkdownViewer'
-import {SxProp} from '../sx'
-import createSlots from '../utils/create-slots'
-import VisuallyHidden from '../_VisuallyHidden'
+import {SxProp} from '../../sx'
+import createSlots from '../../utils/create-slots'
+import VisuallyHidden from '../../_VisuallyHidden'
 import {FormattingTools} from './_FormattingTools'
 import {MarkdownEditorContext} from './_MarkdownEditorContext'
 import {CoreToolbar, DefaultToolbarButtons} from './Toolbar'
