+++ src/drafts/MarkdownEditor/Toolbar.tsx
@@ -16,9 +16,9 @@ import React, {forwardRef, memo, useContext, useRef} from 'react'
 
 import '@github/markdown-toolbar-element'
 import {isMacOS} from '@primer/behaviors/utils'
-import Box from '../Box'
-import {IconButton, IconButtonProps} from '../Button'
-import {useFocusZone} from '../hooks/useFocusZone'
+import Box from '../../Box'
+import {IconButton, IconButtonProps} from '../../Button'
+import {useFocusZone} from '../../hooks/useFocusZone'
 import {MarkdownEditorSlot} from './MarkdownEditor'
 import {MarkdownEditorContext} from './_MarkdownEditorContext'
 import {SavedRepliesButton} from './_SavedReplies'
