+++ src/drafts/MarkdownEditor/_Footer.tsx
@@ -1,7 +1,7 @@
 import React, {memo, useContext} from 'react'
 import {AlertIcon, ImageIcon, MarkdownIcon} from '@primer/octicons-react'
 
-import {Spinner, Button, ButtonProps, LinkButton, Box, Text} from '..'
+import {Spinner, Button, ButtonProps, LinkButton, Box, Text} from '../..'
 import {MarkdownEditorContext} from './_MarkdownEditorContext'
 
 const uploadingNote = ([current, total]: [number, number]) =>
