+++ src/drafts/MarkdownEditor/_MarkdownInput.tsx
@@ -1,9 +1,9 @@
 import {subscribe as subscribeToMarkdownPasting} from '@github/paste-markdown'
 import React, {forwardRef, useEffect, useMemo, useState} from 'react'
-import {useDynamicTextareaHeight} from '../hooks'
-import {useCombinedRefs} from '../hooks/useCombinedRefs'
-import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions} from '../drafts/InlineAutocomplete'
-import Textarea, {TextareaProps} from '../Textarea'
+import {useDynamicTextareaHeight} from '../hooks/useDynamicTextareaHeight'
+import {useCombinedRefs} from '../../hooks/useCombinedRefs'
+import InlineAutocomplete, {ShowSuggestionsEvent, Suggestions} from '../InlineAutocomplete'
+import Textarea, {TextareaProps} from '../../Textarea'
 import {Emoji, useEmojiSuggestions} from './suggestions/_useEmojiSuggestions'
 import {Mentionable, useMentionSuggestions} from './suggestions/_useMentionSuggestions'
 import {Reference, useReferenceSuggestions} from './suggestions/_useReferenceSuggestions'
