+++ src/drafts/MarkdownEditor/_ViewSwitch.tsx
@@ -1,8 +1,8 @@
 import React, {useContext} from 'react'
 import {EyeIcon, PencilIcon} from '@primer/octicons-react'
 
-import Box from '../Box'
-import {Button, IconButton} from '../Button'
+import Box from '../../Box'
+import {Button, IconButton} from '../../Button'
 import {MarkdownEditorContext} from './_MarkdownEditorContext'
 
 export type MarkdownViewMode = 'preview' | 'edit'
