+++ src/drafts/MarkdownEditor/suggestions/_useReferenceSuggestions.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import {suggestionsCalculator, UseSuggestionsHook} from '.'
-import {ActionList} from '../../ActionList'
+import {ActionList} from '../../../ActionList'
 import {Suggestion, Trigger} from '../../InlineAutocomplete'
-import Text from '../../Text'
+import Text from '../../../Text'
 import {score} from 'fzy.js'
 
 export type Reference = {
