+++ src/drafts/hooks/useIgnoreKeyboardActionsWhileComposing.test.tsx
@@ -3,7 +3,7 @@ import {isMacOS} from '@primer/behaviors/utils'
 import {fireEvent, render, screen} from '@testing-library/react'
 import {renderHook} from '@testing-library/react-hooks'
 import userEvent from '@testing-library/user-event'
-import {useIgnoreKeyboardActionsWhileComposing} from '../../hooks/useIgnoreKeyboardActionsWhileComposing'
+import {useIgnoreKeyboardActionsWhileComposing} from './useIgnoreKeyboardActionsWhileComposing'
 
 jest.mock('@primer/behaviors/utils')
