+++ src/hooks/index.ts
@@ -13,6 +13,4 @@ export {useProvidedStateOrCreate} from './useProvidedStateOrCreate'
 export {useMenuInitialFocus} from './useMenuInitialFocus'
 export {useMenuKeyboardNavigation} from './useMenuKeyboardNavigation'
 export {useMnemonics} from './useMnemonics'
-export * from './useUnifiedFileSelect'
-export {useDynamicTextareaHeight} from './useDynamicTextareaHeight'
 export {useCombinedRefs} from './useCombinedRefs'
